tracker->report_last_ms must be u64
It must be u64 as it takes its value from tracker->update_timestamp_ms,
which is u64 value.

+++ src/track/tracker.h
@@ -429,7 +429,7 @@ typedef struct {
   };
 
   /* for tracking profile stabilization timeout bookkeeping */
-  u32 report_last_ms;
+  u64 report_last_ms;
 } tracker_t;
 
 #endif /* SWIFTNAV_TRACKER_H  */
